@@ -14,11 +14,14 @@ description: A new Flutter project.
 version: 1.0.0+1
 
 environment:
-  sdk: ">=2.1.0 <3.0.0"
+  sdk: '>=2.6.0 <3.0.0'
+
 
 dependencies:
   flutter:
     sdk: flutter
+
+
   flutter_localizations:
     sdk: flutter
   flutter_cupertino_localizations: ^1.0.1
@@ -34,15 +37,12 @@ dependencies:
   sentry: '>=3.0.0 <4.0.0'
   transparent_image: ^1.0.0
 
-  
-  just_audio: ^0.1.3
-  flutter_tts: ^0.9.1
 
-
-  # just_audio: ^0.0.6
-  # flutter_tts: ^0.8.5
+  path_provider: ^1.5.1
+  just_audio: ^0.0.6
   rxdart: ^0.23.1
   audio_service: ^0.6.2
+
 
   # The following adds the Cupertino Icons font to your application.
   # Use with the CupertinoIcons class for iOS style icons.
