@@ -1,144 +1,454 @@
-library crashy;
+import 'dart:math';
 
+import 'package:flutter/material.dart';
 import 'dart:async';
 
-import 'package:daff_app/helpers/dsn.dart';
-import 'package:daff_app/screens/splash_screen.dart';
-import 'package:daff_app/widgets/audio_player.dart';
-import 'package:flutter/foundation.dart';
-import 'package:flutter/material.dart';
-import 'package:flutter/services.dart';
-import 'package:sentry/sentry.dart';
-import 'package:daff_app/authentication_model.dart';
-import 'package:daff_app/helpers/firebase_api.dart';
-import 'package:daff_app/providers/author_screen_provider.dart';
-import 'package:daff_app/providers/home_screen_provider.dart';
-import 'package:daff_app/providers/stories_provider.dart';
-import 'package:daff_app/providers/story_screen_provider.dart';
-import 'package:daff_app/screens/home_screen.dart';
-import 'package:google_fonts/google_fonts.dart';
-import 'package:provider/provider.dart';
-import 'package:flutter_localizations/flutter_localizations.dart';
+import 'package:audio_service/audio_service.dart';
+import 'package:just_audio/just_audio.dart';
+import 'package:rxdart/rxdart.dart';
 
-final SentryClient _sentry = new SentryClient(dsn: dsn);
+MediaControl playControl = MediaControl(
+  androidIcon: 'drawable/ic_action_play_arrow',
+  label: 'Play',
+  action: MediaAction.play,
+);
+MediaControl pauseControl = MediaControl(
+  androidIcon: 'drawable/ic_action_pause',
+  label: 'Pause',
+  action: MediaAction.pause,
+);
+MediaControl skipToNextControl = MediaControl(
+  androidIcon: 'drawable/ic_action_skip_next',
+  label: 'Next',
+  action: MediaAction.skipToNext,
+);
+MediaControl skipToPreviousControl = MediaControl(
+  androidIcon: 'drawable/ic_action_skip_previous',
+  label: 'Previous',
+  action: MediaAction.skipToPrevious,
+);
+MediaControl stopControl = MediaControl(
+  androidIcon: 'drawable/ic_action_stop',
+  label: 'Stop',
+  action: MediaAction.stop,
+);
 
-bool get isInDebugMode {
-  bool inDebugMode = false;
-  assert(inDebugMode = true);
-  return inDebugMode;
-}
-
-/// Reports [error] along with its [stackTrace] to Sentry.io.
-Future<Null> _reportError(dynamic error, dynamic stackTrace) async {
-  print('Caught error: $error');
-
-  // Errors thrown in development mode are unlikely to be interesting. You can
-  // check if you are running in dev mode using an assertion and omit sending
-  // the report.
-  if (isInDebugMode) {
-    print(stackTrace);
-    print('In dev mode. Not sending report to Sentry.io.');
-    return;
-  }
-
-  print('Reporting to Sentry.io...');
-
-  final SentryResponse response = await _sentry.captureException(
-    exception: error,
-    stackTrace: stackTrace,
-  );
-
-  if (response.isSuccessful) {
-    print('Success! Event ID: ${response.eventId}');
-  } else {
-    print('Failed to report to Sentry.io: ${response.error}');
-  }
-}
-
-Future<Null> main() async {
-  // This captures errors reported by the Flutter framework.
-  FlutterError.onError = (FlutterErrorDetails details) async {
-    if (isInDebugMode) {
-      // In development mode simply print to console.
-      FlutterError.dumpErrorToConsole(details);
-    } else {
-      // In production mode report to the application zone to report to
-      // Sentry.
-      Zone.current.handleUncaughtError(details.exception, details.stack);
-    }
-  };
-
-  runZoned<Future<Null>>(() async {
-    // runApp(new MyApp());
-    runApp(MyAudioApp());
-  }, onError: (error, stackTrace) async {
-    await _reportError(error, stackTrace);
-  });
-}
-
-
+void main() => runApp(new MyApp());
 
 class MyApp extends StatefulWidget {
-  MyApp();
   @override
-  _MyAppState createState() => _MyAppState();
+  _MyAppState createState() => new _MyAppState();
 }
 
-class _MyAppState extends State<MyApp> {
-  FirebaseAPI firebaseAPI;
-  HomeModel homeModel;
-  StoriesModel storiesModel ;
+class _MyAppState extends State<MyApp> with WidgetsBindingObserver {
+  final BehaviorSubject<double> _dragPositionSubject =
+      BehaviorSubject.seeded(null);
 
+  @override
   void initState() {
-    SystemChrome.setSystemUIOverlayStyle(SystemUiOverlayStyle(
-      statusBarColor: Colors.white,
-    ));
-
     super.initState();
-    firebaseAPI = FirebaseAPI();
-    firebaseAPI.initialize();
-    storiesModel = StoriesModel();
-    // homeModel = HomeModel();
-    // homeModel.initialize();
-    
+    WidgetsBinding.instance.addObserver(this);
+    connect();
   }
 
+  @override
+  void dispose() {
+    disconnect();
+    WidgetsBinding.instance.removeObserver(this);
+    super.dispose();
+  }
 
-  _MyAppState();
+  @override
+  void didChangeAppLifecycleState(AppLifecycleState state) {
+    switch (state) {
+      case AppLifecycleState.resumed:
+        connect();
+        break;
+      case AppLifecycleState.paused:
+        disconnect();
+        break;
+      default:
+        break;
+    }
+  }
+
+  void connect() async {
+    await AudioService.connect();
+  }
+
+  void disconnect() {
+    AudioService.disconnect();
+  }
+
   @override
   Widget build(BuildContext context) {
-    return MultiProvider(
-      providers: [
-        ChangeNotifierProvider(create: (_) => AuthenticationModel()),
-        ChangeNotifierProvider(create: (_) => firebaseAPI),
-        ChangeNotifierProvider(create: (_) => homeModel),
-        ChangeNotifierProvider(create: (_) => StoryModel()),
-        ChangeNotifierProvider(create: (_) =>  storiesModel),
-        ChangeNotifierProvider(create: (_) =>  AuthorModel()),
-      ],
-  child: MaterialApp(
+    return new MaterialApp(
+      home: WillPopScope(
+        onWillPop: () {
+          disconnect();
+          return Future.value(true);
+        },
+        child: new Scaffold(
+          appBar: new AppBar(
+            title: const Text('Audio Service Demo'),
+          ),
+          body: new Center(
+            child: StreamBuilder<ScreenState>(
+              stream: Rx.combineLatest3<List<MediaItem>, MediaItem,
+                      PlaybackState, ScreenState>(
+                  AudioService.queueStream,
+                  AudioService.currentMediaItemStream,
+                  AudioService.playbackStateStream,
+                  (queue, mediaItem, playbackState) =>
+                      ScreenState(queue, mediaItem, playbackState)),
+              builder: (context, snapshot) {
+                final screenState = snapshot.data;
+                final queue = screenState?.queue;
+                final mediaItem = screenState?.mediaItem;
+                final state = screenState?.playbackState;
+                final basicState = state?.basicState ?? BasicPlaybackState.none;
+                return Column(
+                  mainAxisAlignment: MainAxisAlignment.center,
+                  children: [
+                    if (queue != null && queue.isNotEmpty)
+                      Row(
+                        mainAxisAlignment: MainAxisAlignment.center,
+                        children: [
+                          IconButton(
+                            icon: Icon(Icons.skip_previous),
+                            iconSize: 64.0,
+                            onPressed: mediaItem == queue.first
+                                ? null
+                                : AudioService.skipToPrevious,
+                          ),
+                          IconButton(
+                            icon: Icon(Icons.skip_next),
+                            iconSize: 64.0,
+                            onPressed: mediaItem == queue.last
+                                ? null
+                                : AudioService.skipToNext,
+                          ),
+                        ],
+                      ),
+                    if (mediaItem?.title != null) Text(mediaItem.title),
+                    if (basicState == BasicPlaybackState.none) ...[
+                      audioPlayerButton(),
+                    ] else
+                      Row(
+                        mainAxisAlignment: MainAxisAlignment.center,
+                        children: [
+                          if (basicState == BasicPlaybackState.playing)
+                            pauseButton()
+                          else if (basicState == BasicPlaybackState.paused)
+                            playButton()
+                          else if (basicState == BasicPlaybackState.buffering ||
+                              basicState == BasicPlaybackState.skippingToNext ||
+                              basicState ==
+                                  BasicPlaybackState.skippingToPrevious)
+                            Padding(
+                              padding: const EdgeInsets.all(8.0),
+                              child: SizedBox(
+                                width: 64.0,
+                                height: 64.0,
+                                child: CircularProgressIndicator(),
+                              ),
+                            ),
+                          stopButton(),
+                        ],
+                      ),
+                    if (basicState != BasicPlaybackState.none &&
+                        basicState != BasicPlaybackState.stopped) ...[
+                      positionIndicator(mediaItem, state),
+                      Text("State: " +
+                          "$basicState".replaceAll(RegExp(r'^.*\.'), '')),
+                    ]
+                  ],
+                );
+              },
+            ),
+          ),
+        ),
+      ),
+    );
+  }
 
-    localizationsDelegates: [
-      GlobalMaterialLocalizations.delegate,
-      GlobalWidgetsLocalizations.delegate,
-      // GlobalCupertinoLocalizations.delegate,
-    ],
-    supportedLocales: [
-      const Locale('he'), // Hebrew
-    ],
-    locale: Locale("he", "HE") ,
-    title: 'Daff Rocking App',
-    theme: ThemeData(
-      primarySwatch: Colors.grey,
-      textTheme: GoogleFonts.alefTextTheme(Theme.of(context).textTheme,),
-    ),
-    home: Directionality( // add this
-      textDirection: TextDirection.rtl, // set this property 
-      child: SplashScreen(),
-    ),
-    routes: {
-      HomeScreen.routeName: (ctx) => HomeScreen(),
-    },
-  ));
+  RaisedButton audioPlayerButton() => startButton(
+        'AudioPlayer',
+        () {
+          AudioService.start(
+            backgroundTaskEntrypoint: _audioPlayerTaskEntrypoint,
+            androidNotificationChannelName: 'Audio Service Demo',
+            notificationColor: 0xFF2196f3,
+            androidNotificationIcon: 'mipmap/ic_launcher',
+            enableQueue: true,
+          );
+        },
+      );
+
+
+
+  RaisedButton startButton(String label, VoidCallback onPressed) =>
+      RaisedButton(
+        child: Text(label),
+        onPressed: onPressed,
+      );
+
+  IconButton playButton() => IconButton(
+        icon: Icon(Icons.play_arrow),
+        iconSize: 64.0,
+        onPressed: AudioService.play,
+      );
+
+  IconButton pauseButton() => IconButton(
+        icon: Icon(Icons.pause),
+        iconSize: 64.0,
+        onPressed: AudioService.pause,
+      );
+
+  IconButton stopButton() => IconButton(
+        icon: Icon(Icons.stop),
+        iconSize: 64.0,
+        onPressed: AudioService.stop,
+      );
+
+  Widget positionIndicator(MediaItem mediaItem, PlaybackState state) {
+    double seekPos;
+    return StreamBuilder(
+      stream: Rx.combineLatest2<double, double, double>(
+          _dragPositionSubject.stream,
+          Stream.periodic(Duration(milliseconds: 200)),
+          (dragPosition, _) => dragPosition),
+      builder: (context, snapshot) {
+        double position = snapshot.data ?? state.currentPosition.toDouble();
+        double duration = mediaItem?.duration?.toDouble();
+        return Column(
+          children: [
+            if (duration != null)
+              Slider(
+                min: 0.0,
+                max: duration,
+                value: seekPos ?? max(0.0, min(position, duration)),
+                onChanged: (value) {
+                  _dragPositionSubject.add(value);
+                },
+                onChangeEnd: (value) {
+                  AudioService.seekTo(value.toInt());
+                  // Due to a delay in platform channel communication, there is
+                  // a brief moment after releasing the Slider thumb before the
+                  // new position is broadcast from the platform side. This
+                  // hack is to hold onto seekPos until the next state update
+                  // comes through.
+                  // TODO: Improve this code.
+                  seekPos = value;
+                  _dragPositionSubject.add(null);
+                },
+              ),
+            Text("${(state.currentPosition / 1000).toStringAsFixed(3)}"),
+          ],
+        );
+      },
+    );
   }
 }
+
+class ScreenState {
+  final List<MediaItem> queue;
+  final MediaItem mediaItem;
+  final PlaybackState playbackState;
+
+  ScreenState(this.queue, this.mediaItem, this.playbackState);
+}
+
+void _audioPlayerTaskEntrypoint() async {
+  AudioServiceBackground.run(() => AudioPlayerTask());
+}
+
+class AudioPlayerTask extends BackgroundAudioTask {
+  final _queue = <MediaItem>[
+    MediaItem(
+      id: "https://s3.amazonaws.com/scifri-episodes/scifri20181123-episode.mp3",
+      album: "Science Friday",
+      title: "A Salute To Head-Scratching Science",
+      artist: "Science Friday and WNYC Studios",
+      duration: 5739820,
+      artUri:
+          "https://media.wnyc.org/i/1400/1400/l/80/1/ScienceFriday_WNYCStudios_1400.jpg",
+    ),
+    MediaItem(
+      id: "https://s3.amazonaws.com/scifri-segments/scifri201711241.mp3",
+      album: "Science Friday",
+      title: "From Cat Rheology To Operatic Incompetence",
+      artist: "Science Friday and WNYC Studios",
+      duration: 2856950,
+      artUri:
+          "https://media.wnyc.org/i/1400/1400/l/80/1/ScienceFriday_WNYCStudios_1400.jpg",
+    ),
+  ];
+  int _queueIndex = -1;
+  AudioPlayer _audioPlayer = new AudioPlayer();
+  Completer _completer = Completer();
+  BasicPlaybackState _skipState;
+  bool _playing;
+
+  bool get hasNext => _queueIndex + 1 < _queue.length;
+
+  bool get hasPrevious => _queueIndex > 0;
+
+  MediaItem get mediaItem => _queue[_queueIndex];
+
+  BasicPlaybackState _stateToBasicState(AudioPlaybackState state) {
+    switch (state) {
+      case AudioPlaybackState.none:
+        return BasicPlaybackState.none;
+      case AudioPlaybackState.stopped:
+        return BasicPlaybackState.stopped;
+      case AudioPlaybackState.paused:
+        return BasicPlaybackState.paused;
+      case AudioPlaybackState.playing:
+        return BasicPlaybackState.playing;
+      case AudioPlaybackState.buffering:
+        return BasicPlaybackState.buffering;
+      case AudioPlaybackState.connecting:
+        return _skipState ?? BasicPlaybackState.connecting;
+      case AudioPlaybackState.completed:
+        return BasicPlaybackState.stopped;
+      default:
+        throw Exception("Illegal state");
+    }
+  }
+
+  @override
+  Future<void> onStart() async {
+    var playerStateSubscription = _audioPlayer.playbackStateStream
+        .where((state) => state == AudioPlaybackState.completed)
+        .listen((state) {
+      _handlePlaybackCompleted();
+    });
+    var eventSubscription = _audioPlayer.playbackEventStream.listen((event) {
+      final state = _stateToBasicState(event.state);
+      if (state != BasicPlaybackState.stopped) {
+        _setState(
+          state: state,
+          position: event.position.inMilliseconds,
+        );
+      }
+    });
+
+    AudioServiceBackground.setQueue(_queue);
+    await onSkipToNext();
+    await _completer.future;
+    playerStateSubscription.cancel();
+    eventSubscription.cancel();
+  }
+
+  void _handlePlaybackCompleted() {
+    if (hasNext) {
+      onSkipToNext();
+    } else {
+      onStop();
+    }
+  }
+
+  void playPause() {
+    if (AudioServiceBackground.state.basicState == BasicPlaybackState.playing)
+      onPause();
+    else
+      onPlay();
+  }
+
+  @override
+  Future<void> onSkipToNext() => _skip(1);
+
+  @override
+  Future<void> onSkipToPrevious() => _skip(-1);
+
+  Future<void> _skip(int offset) async {
+    final newPos = _queueIndex + offset;
+    if (!(newPos >= 0 && newPos < _queue.length)) return;
+    if (_playing == null) {
+      // First time, we want to start playing
+      _playing = true;
+    } else if (_playing) {
+      // Stop current item
+      await _audioPlayer.stop();
+    }
+    // Load next item
+    _queueIndex = newPos;
+    AudioServiceBackground.setMediaItem(mediaItem);
+    _skipState = offset > 0
+        ? BasicPlaybackState.skippingToNext
+        : BasicPlaybackState.skippingToPrevious;
+    await _audioPlayer.setUrl(mediaItem.id);
+    _skipState = null;
+    // Resume playback if we were playing
+    if (_playing) {
+      onPlay();
+    } else {
+      _setState(state: BasicPlaybackState.paused);
+    }
+  }
+
+  @override
+  void onPlay() {
+    if (_skipState == null) {
+      _playing = true;
+      _audioPlayer.play();
+    }
+  }
+
+  @override
+  void onPause() {
+    if (_skipState == null) {
+      _playing = false;
+      _audioPlayer.pause();
+    }
+  }
+
+  @override
+  void onSeekTo(int position) {
+    _audioPlayer.seek(Duration(milliseconds: position));
+  }
+
+  @override
+  void onClick(MediaButton button) {
+    playPause();
+  }
+
+  @override
+  void onStop() {
+    _audioPlayer.stop();
+    _setState(state: BasicPlaybackState.stopped);
+    _completer.complete();
+  }
+
+  void _setState({@required BasicPlaybackState state, int position}) {
+    if (position == null) {
+      position = _audioPlayer.playbackEvent.position.inMilliseconds;
+    }
+    AudioServiceBackground.setState(
+      controls: getControls(state),
+      systemActions: [MediaAction.seekTo],
+      basicState: state,
+      position: position,
+    );
+  }
+
+  List<MediaControl> getControls(BasicPlaybackState state) {
+    if (_playing) {
+      return [
+        skipToPreviousControl,
+        pauseControl,
+        stopControl,
+        skipToNextControl
+      ];
+    } else {
+      return [
+        skipToPreviousControl,
+        playControl,
+        stopControl,
+        skipToNextControl
+      ];
+    }
+  }
+}
+
